🚨 Fix compilation of double function displayImage
