+++ src/Draw/draw.cpp
@@ -25,78 +25,6 @@ static double getLuminescence(RayTracer::Point3D &intersection,
     return Light->getLuminescence(distance, angle);
 }
 
-void createPPMFile(const sf::Image& image,
-    const std::string& filename) {
-    unsigned int width;
-    unsigned int height;
-    std::ofstream out(filename);
-
-    width = image.getSize().x;
-    height = image.getSize().y;
-    out << "P3\n" << width << " " << height << "\n255\n";
-    for (unsigned int y = 0; y < height; ++y) {
-        for (unsigned int x = 0; x < width; ++x) {
-            sf::Color color = image.getPixel(x, y);
-            out << static_cast<int>(color.r) << " "
-                << static_cast<int>(color.g) << " "
-                << static_cast<int>(color.b) << "  ";
-        }
-        out << "\n";
-    }
-    out.close();
-}
-
-std::string randomString(const int len) {
-    static const char alphanum[] =
-        "0123456789"
-        "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-        "abcdefghijklmnopqrstuvwxyz";
-    std::string tmp_s;
-
-    tmp_s.reserve(len);
-    for (int i = 0; i < len; ++i)
-        tmp_s += alphanum[rand() % (sizeof(alphanum) - 1)];
-    return ("renders/screenshot" + tmp_s + ".ppm");
-}
-
-void renderAtClosing(sf::RenderWindow &window) {
-    sf::Texture texture;
-    sf::Image screenshot;
-
-    texture.create(window.getSize().x, window.getSize().y);
-    texture.update(window);
-    screenshot  = texture.copyToImage();
-    createPPMFile(screenshot, "renders/render.ppm");
-}
-
-void screenshot(sf::RenderWindow &window) {
-    sf::Texture texture;
-    sf::Image screenshot;
-
-    texture.create(window.getSize().x, window.getSize().y);
-    texture.update(window);
-    screenshot  = texture.copyToImage();
-    createPPMFile(screenshot, randomString(5));
-}
-
-void displayImage(sf::RenderWindow &window, sf::Image &image) {
-    while (window.isOpen()) {
-        sf::Event event;
-        while (window.pollEvent(event)) {
-            if (event.type == sf::Event::Closed
-                || event.key.code == sf::Keyboard::Escape) {
-                renderAtClosing(window);
-                window.close();
-            }
-            if (event.type == sf::Event::KeyPressed
-                && event.key.code == sf::Keyboard::Space) {
-                screenshot(window);
-            }
-        }
-        showImage(window, image);
-    }
-}
-
 static void hit(sf::Image &image, int i, int j, RayTracer::Ray &ray,
     std::shared_ptr<Prim> &s, RayTracer::Point3D &intersection,
     std::unique_ptr<Light> &Light, float &minRayLen) {
